Api change for mm
