+++ helix-rest/src/main/java/org/apache/helix/rest/server/service/InstanceService.java
@@ -20,63 +20,14 @@ package org.apache.helix.rest.server.service;
  */
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
-import com.google.common.collect.ImmutableList;
+import org.apache.helix.rest.clusterMaintenanceService.HealthCheck;
 import org.apache.helix.rest.server.json.instance.InstanceInfo;
 import org.apache.helix.rest.server.json.instance.StoppableCheck;
 
 public interface InstanceService {
-    enum HealthCheck {
-        /**
-         * Check if instance is alive
-         */
-        INSTANCE_NOT_ALIVE,
-        /**
-         * Check if instance is enabled both in instance config and cluster config
-         */
-        INSTANCE_NOT_ENABLED,
-        /**
-         * Check if instance is stable
-         * Stable means all the ideal state mapping matches external view (view of current state).
-         */
-        INSTANCE_NOT_STABLE,
-        /**
-         * Check if instance has 0 resource assigned
-         */
-        EMPTY_RESOURCE_ASSIGNMENT,
-        /**
-         * Check if instance has disabled partitions
-         */
-        HAS_DISABLED_PARTITION,
-        /**
-         * Check if instance has valid configuration (pre-requisite for all checks)
-         */
-        INVALID_CONFIG,
-        /**
-         * Check if instance has error partitions
-         */
-        HAS_ERROR_PARTITION,
-        /**
-         * Check if all resources hosted on the instance can still meet the min active replica
-         * constraint if this instance is shutdown
-         */
-        MIN_ACTIVE_REPLICA_CHECK_FAILED;
-
-        /**
-         * Pre-defined list of checks to test if an instance can be stopped at runtime
-         */
-        public static List<HealthCheck> STOPPABLE_CHECK_LIST = Arrays.asList(HealthCheck.values());
-        /**
-         * Pre-defined list of checks to test if an instance is in healthy running state
-         */
-        public static List<HealthCheck> STARTED_AND_HEALTH_CHECK_LIST = ImmutableList
-            .of(INVALID_CONFIG, INSTANCE_NOT_ALIVE, INSTANCE_NOT_ENABLED, INSTANCE_NOT_STABLE,
-                EMPTY_RESOURCE_ASSIGNMENT);
-    }
-
     /**
      * Get the overall status of the instance
      *
@@ -84,6 +35,7 @@ public interface InstanceService {
      * @param instanceName The instance name
      * @return An instance of {@link InstanceInfo} easily convertible to JSON
      */
+    @Deprecated
     InstanceInfo getInstanceInfo(String clusterId, String instanceName,
                                  List<HealthCheck> healthChecks);
 
@@ -96,6 +48,7 @@ public interface InstanceService {
      * @return An instance of {@link StoppableCheck} easily convertible to JSON
      * @throws IOException in case of network failure
      */
+    @Deprecated
     StoppableCheck getInstanceStoppableCheck(String clusterId, String instanceName,
                                              String jsonContent) throws IOException;
 
@@ -108,6 +61,7 @@ public interface InstanceService {
      * @return A map contains the instance as key and the StoppableCheck as the value
      * @throws IOException in case of network failure
      */
+    @Deprecated
     Map<String, StoppableCheck> batchGetInstancesStoppableChecks(String clusterId, List<String> instances, String jsonContent)
             throws IOException;
 }
