+++ helix-rest/src/test/java/org/apache/helix/rest/clusterMaintenanceService/TestMaintenanceManagementService.java
@@ -1,4 +1,4 @@
-package org.apache.helix.rest.server.service;
+package org.apache.helix.rest.clusterMaintenanceService;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -57,10 +57,11 @@ import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyZeroInteractions;
+import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
-public class TestInstanceService {
+
+public class TestMaintenanceManagementService {
   private static final String TEST_CLUSTER = "TestCluster";
   private static final String TEST_INSTANCE = "instance0.linkedin.com_1235";
 
@@ -79,12 +80,26 @@ public class TestInstanceService {
     when(_configAccessor.getRESTConfig(TEST_CLUSTER)).thenReturn(restConfig);
   }
 
+  class MockMaintenanceManagementService extends MaintenanceManagementService {
+    public MockMaintenanceManagementService(ZKHelixDataAccessor dataAccessor,
+        ConfigAccessor configAccessor, CustomRestClient customRestClient, boolean skipZKRead,
+        boolean continueOnFailures, String namespace) {
+      super(dataAccessor, configAccessor, customRestClient, skipZKRead, continueOnFailures,
+          namespace);
+    }
+    @Override
+    protected Map<String, Boolean> getInstanceHealthStatus(String clusterId, String instanceName,
+        List<HealthCheck> healthChecks) {
+      return Collections.emptyMap();
+    }
+  }
+
   @Test
   public void testGetInstanceStoppableCheckWhenHelixOwnCheckFail() throws IOException {
     Map<String, Boolean> failedCheck = ImmutableMap.of("FailCheck", false);
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
-            HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient,
+            false, false, HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
               String instanceName, List<HealthCheck> healthChecks) {
@@ -98,14 +113,14 @@ public class TestInstanceService {
 
     Assert.assertEquals(actual.getFailedChecks().size(), failedCheck.size());
     Assert.assertFalse(actual.isStoppable());
-    verifyZeroInteractions(_customRestClient);
-    verifyZeroInteractions(_configAccessor);
+    verifyNoInteractions(_customRestClient);
+    verifyNoInteractions(_configAccessor);
   }
 
   @Test
   public void testGetInstanceStoppableCheckWhenCustomInstanceCheckFail() throws IOException {
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient, false, false,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -129,8 +144,8 @@ public class TestInstanceService {
 
   @Test
   public void testGetInstanceStoppableCheckConnectionRefused() throws IOException {
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient, false, false,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -210,17 +225,17 @@ public class TestInstanceService {
         AccessOption.PERSISTENT)).thenReturn(MasterSlaveSMD.build().getRecord());
 
     // Valid data only from ZK, pass the check
-    InstanceService instanceServiceReadZK =
-        new MockInstanceServiceImpl(zkHelixDataAccessor, _configAccessor, _customRestClient, false,
-            false);
+    MockMaintenanceManagementService instanceServiceReadZK =
+        new MockMaintenanceManagementService(zkHelixDataAccessor, _configAccessor, _customRestClient, false,
+            false, HelixRestNamespace.DEFAULT_NAMESPACE_NAME);
     StoppableCheck stoppableCheck =
         instanceServiceReadZK.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
     Assert.assertTrue(stoppableCheck.isStoppable());
 
     // Even ZK data is valid. Skip ZK read should fail the test.
-    InstanceService instanceServiceWithoutReadZK =
-        new MockInstanceServiceImpl(zkHelixDataAccessor, _configAccessor, _customRestClient, true,
-            false);
+    MockMaintenanceManagementService instanceServiceWithoutReadZK =
+        new MockMaintenanceManagementService(zkHelixDataAccessor, _configAccessor, _customRestClient, true,
+            false, HelixRestNamespace.DEFAULT_NAMESPACE_NAME);
     stoppableCheck = instanceServiceWithoutReadZK.getInstanceStoppableCheck(TEST_CLUSTER, TEST_INSTANCE, jsonContent);
     Assert.assertFalse(stoppableCheck.isStoppable());
   }
@@ -240,8 +255,8 @@ public class TestInstanceService {
         .thenReturn(Arrays.asList(TEST_INSTANCE, siblingInstance));
 
     Map<String, Boolean> instanceHealthFailedCheck = ImmutableMap.of("FailCheck", false);
-    InstanceService service =
-        new InstanceServiceImpl(zkHelixDataAccessor, _configAccessor, _customRestClient, true, true,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(zkHelixDataAccessor, _configAccessor, _customRestClient, true, true,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -272,8 +287,8 @@ public class TestInstanceService {
   // TODO re-enable the test when partition health checks get decoupled
   @Test(enabled = false)
   public void testGetInstanceStoppableCheckWhenPartitionsCheckFail() throws IOException {
-    InstanceService service =
-        new InstanceServiceImpl(_dataAccessor, _configAccessor, _customRestClient, false, false,
+    MockMaintenanceManagementService service =
+        new MockMaintenanceManagementService(_dataAccessor, _configAccessor, _customRestClient, false, false,
             HelixRestNamespace.DEFAULT_NAMESPACE_NAME) {
           @Override
           protected Map<String, Boolean> getInstanceHealthStatus(String clusterId,
@@ -293,17 +308,4 @@ public class TestInstanceService {
     verify(_customRestClient, times(1)).getInstanceStoppableCheck(any(), any());
   }
 
-  private static class MockInstanceServiceImpl extends InstanceServiceImpl {
-    MockInstanceServiceImpl(ZKHelixDataAccessor dataAccessor, ConfigAccessor configAccessor,
-        CustomRestClient customRestClient, boolean skipZKRead, boolean continueOnFailures) {
-      super(dataAccessor, configAccessor, customRestClient, skipZKRead, continueOnFailures,
-          HelixRestNamespace.DEFAULT_NAMESPACE_NAME);
-    }
-
-    @Override
-    protected Map<String, Boolean> getInstanceHealthStatus(String clusterId, String instanceName,
-        List<InstanceService.HealthCheck> healthChecks) {
-      return Collections.emptyMap();
-    }
-  }
 }
